(API) Setting up Cookie authentication

+++ apps/api/src/token/token.service.ts
@@ -7,6 +7,8 @@ import { PrismaService } from '@/prisma/prisma.service'
 import { Cron, CronExpression } from '@nestjs/schedule'
 import { type Token, type User } from '@prisma/client'
 import { type Request } from 'express'
+import * as ms from 'ms'
+import { type StringValue } from 'ms'
 
 @Injectable()
 export class TokenService extends PassportStrategy(Strategy) {
@@ -33,21 +35,39 @@ export class TokenService extends PassportStrategy(Strategy) {
         return { userId: payload.sub, username: payload.username }
     }
 
+    public async isRegistered(token: string): Promise<boolean> {
+        return (
+            (await this.prismaService.token.findUnique({
+                where: { token }
+            })) != null
+        )
+    }
+
     public async create(user: User): Promise<Token> {
         const token = await this.prismaService.token.create({
             data: {
                 token: this.jwtService.sign({ sub: user.id }),
                 userId: user.id,
-                expiresAt: new Date(new Date().getTime() + 60 * 60 * 1000)
+                expiresAt: new Date(
+                    Date.now() +
+                        ms(
+                            this.configService.get<StringValue>(
+                                'JWT_EXPIRES_IN',
+                                '1h'
+                            )
+                        )
+                )
             }
         })
         return token
     }
 
-    public async deleteMany(user: User): Promise<void> {
-        await this.prismaService.token.deleteMany({
-            where: { userId: user.id }
-        })
+    public async deleteMany(user: User | null): Promise<void> {
+        if (user != null) {
+            await this.prismaService.token.deleteMany({
+                where: { userId: user.id }
+            })
+        }
     }
 
     @Cron(CronExpression.EVERY_HOUR)
